Revert "Async refactoring"

+++ src/costs/services.py
@@ -1,31 +1,18 @@
 import calendar
 from datetime import date, datetime
-from decimal import Decimal
+from decimal import Decimal, InvalidOperation
 from itertools import groupby
 from operator import attrgetter
-from typing import Iterable, Optional
+from typing import Any, Iterable, Optional
 
 from categories import CategoriesService, Category
 from configurations import ConfigurationsService
-from costs.domain import Cost, CostsError, CostsStorage
+from costs.domain import Cost, CostsError
 from db import database
 from equity import EquityCRUD
 from finances.domain import Currencies, Operations
 from shared.formatting import get_number_in_frames
 from users import User, UsersCRUD
-
-
-class CostsCRUD:
-    __TABLE = "costs"
-
-    @classmethod
-    def get_by_id(cls, cost_id: str) -> Optional[Cost]:
-        data: Optional[dict] = database.fetch(cls.__TABLE, "id", cost_id)
-        return Cost(**data) if data else None
-
-    @classmethod
-    def delete_by_id(cls, cost_id: str) -> None:
-        database.delete(cls.__TABLE, "id", cost_id)
 
 
 class CostsService:
@@ -40,23 +27,54 @@ class CostsService:
         self._text: Optional[str] = None
         self._value: Optional[Decimal] = None
 
-    @classmethod
-    def save_costs(cls, storage: CostsStorage) -> Cost:
-        if not all((storage.description, storage.date, storage.category, storage.value)):
+    def set_category(self, text: Optional[str] = None) -> None:
+        if text is None:
+            raise CostsError("Category is not selected")
+
+        if text not in {c.name for c in CategoriesService.CACHED_CATEGORIES}:
+            raise CostsError("Category is not allowed")
+
+        self._category = CategoriesService.get_by_name(text)
+
+    def set_date(self, text: Optional[str] = None) -> None:
+        if text is None:
+            raise CostsError("Date is not selected")
+
+        try:
+            self._date = datetime.strptime(text, self.__FULL_DATE_FROAMT)
+        except ValueError:
+            raise CostsError("Invalid date")
+
+    def add_text(self, text: Optional[str] = None) -> None:
+        if text is None:
+            raise CostsError("Text is not added")
+
+        self._text = text
+
+    def add_value(self, text: Optional[str] = None) -> None:
+        if text is None:
+            raise CostsError("Value is not added")
+
+        try:
+            self._value = Decimal(text.replace(",", "."))
+        except InvalidOperation:
+            raise CostsError("Money value is invalid")
+
+    def save_costs(self) -> Cost:
+        if not self._text or not self._date or not self._category or not self._user:
             raise CostsError("One or more mandatory values are not set")
 
-        user: User = UsersCRUD.fetch_by_account_id(storage.account_id)
         default_currency: str = ConfigurationsService.get_by_name("default_currency").value
 
         payload = {
-            "name": storage.description,
-            "value": storage.value,
+            "name": self._text,
+            "value": self._value,
             "currency": default_currency,
-            "date": storage.date,
-            "user_id": user.id,
-            "category_id": storage.category.id,  # type: ignore
+            "date": self._date,
+            "user_id": self._user.id,
+            "category_id": self._category.id,
         }
-        data: dict = database.insert(cls.__TABLE, payload)
+        data: dict = database.insert(self.__TABLE, payload)
         cost = Cost(**data)
 
         EquityCRUD.update(
@@ -67,8 +85,15 @@ class CostsService:
 
         return cost
 
-    @staticmethod
-    def get_formatted_cost(cost: Cost) -> str:
+    def process_confirmation(self, text: Optional[str]) -> bool:
+        if not text or "Yes" not in text:
+            return False
+
+        self.save_costs()
+        return True
+
+    @classmethod
+    def get_formatted_cost(cls, cost: Cost) -> str:
         category: Optional[Category] = CategoriesService.get_by_id(cost.category_id)
         user = UsersCRUD.fetch_by_id(cost.user_id)
 
@@ -89,8 +114,8 @@ class CostsService:
     def get_formatted_costs_for_delete(cls, costs: list[Cost]) -> str:
         return "\n".join(cls.get_formatted_cost(cost) for cost in costs)
 
-    @staticmethod
-    def get_costs_by_currency(costs: list[Cost]) -> Iterable:
+    @classmethod
+    def get_costs_by_currency(cls, costs: list[Cost]) -> Iterable:
         attr = "currency"
         return groupby(sorted(costs, key=attrgetter(attr)), key=attrgetter(attr))
 
@@ -144,9 +169,13 @@ class CostsService:
         return {currency: list(costs_iter) for currency, costs_iter in cls.get_costs_by_currency(costs)}
 
     @classmethod
-    def get_by_id(cls, cost_id: str) -> None:
-        database.delete(cls.__TABLE, "id", cost_id)
+    def delete_by_id(cls, cost_id: Any, allowed_ids: set[int]) -> None:
+        try:
+            cost_id = int(cost_id)
+        except ValueError:
+            raise CostsError("Invalid id value. It should be a number")
 
-    @classmethod
-    def delete_by_id(cls, cost_id: str) -> None:
+        if cost_id not in allowed_ids:
+            raise CostsError(f"Can't find id 👉 {cost_id}\nPlease use the keyboard below!")
+
         database.delete(cls.__TABLE, "id", cost_id)
